Move function definition
